Add new files for orca metric refactor.
