+++ src/http_server.cc
@@ -87,16 +87,6 @@ namespace triton { namespace server {
     }                                                                   \
   } while (false)
 
-#define ENDPOINT_LOAD_METRICS_TYPE "endpoint-load-metrics-format"
-#define ENDPOINT_LOAD_METRICS_NAME "endpoint-load-metrics"
-#define KV_CACHE_BLOCK_METRICS_FAMILY "nv_trt_llm_kv_cache_block_metrics"
-#define KV_CACHE_BLOCK_TYPE "kv_cache_block_type"
-#define KV_CACHE_BLOCK_TYPE_TOKENS_PER "tokens_per"
-#define KV_CACHE_BLOCK_TYPE_USED "used"
-#define KV_CACHE_BLOCK_TYPE_MAX "max"
-#define KV_CACHE_UTIL_KEY "kv_cache_utilization"
-#define MAX_TOKEN_CAPACITY_KEY "max_token_capacity"
-#define NAMED_METRICS "named_metrics"
 
 namespace {
 
@@ -4320,174 +4310,6 @@ HTTPAPIServer::GenerateRequestClass::StartResponse(
   evhtp_send_reply_chunk_start(req, infer_request->response_code_);
   evhtp_request_resume(req);
 }
-
-
-#ifdef TRITON_ENABLE_METRICS
-std::vector<HTTPAPIServer::GenerateRequestClass::PromMetric>
-HTTPAPIServer::GenerateRequestClass::MetricFamilyExtractor(
-    const std::string& input, const std::string& metricFamily)
-{
-  std::vector<PromMetric> metrics;
-  // Construct the regex pattern using the provided metricFamily.
-
-  // `labelGroup` is a capturing group that captures all characters within curly
-  // braces, excluding line breaks.
-  std::string labelGroup = "(?:{(.*?)})";
-
-  // `valueGroup` is a capturing group that captures a number with its
-  // decimals if any.
-  std::string valueGroup = R"((\d+(?:\.\d+)?))";
-
-  // `patternStr` matches on lines starting with `metricFamily` then captures
-  // its labels if any, then (optionally) matches any whitespace, then captures
-  // its numeric double value.
-  //
-  // For example, `patternStr` would match on input:
-  // `nv_trt_llm_kv_cache_block_metrics{kv_cache_block_type="used",model="tensorrt_llm",version="1"}
-  // 3`
-  //
-  // with 2 capturing groups:
-  // 1. `kv_cache_block_type="used",model="tensorrt_llm",version="1"`
-  // 2. `3`
-  std::string patternStr = metricFamily + labelGroup + R"(?\s*)" + valueGroup;
-  re2::RE2 pattern(patternStr);
-  re2::StringPiece inputPiece(input);
-
-  std::string labelString;
-  std::string metric_value;
-
-  while (re2::RE2::FindAndConsume(
-      &inputPiece, pattern, &labelString, &metric_value)) {
-    PromMetric metric;
-
-    // Extract labels if they exist
-    if (!labelString.empty()) {
-      // `labelPattern` captures any alphanumeric sequence that precedes an '='
-      // character, then captures the following quoted character sequence. These
-      // groups are exahstive given the prometheus data model:
-      // https://prometheus.io/docs/concepts/data_model/#metric-names-and-labels
-      //
-      // For example, calling FindAndConsume() with `labelPattern` on input:
-      // `kv_cache_block_type="used",model="tensorrt_llm",version="1"`
-      //
-      // matches 3 times with 2 capturing groups each:
-      //
-      // Match #1
-      // 1. `kv_cache_block_type`
-      // 2. `used`
-      //
-      // Match #2
-      // 1. `model`
-      // 2. `tensorrt_llm`
-      //
-      // Match #3
-      // 1. `version`
-      // 2. `1`
-      re2::RE2 labelPattern(R"((\w+)=\"([^\"]*)\")");
-      re2::StringPiece labelPiece(labelString);
-      std::string key, value;
-      while (
-          re2::RE2::FindAndConsume(&labelPiece, labelPattern, &key, &value)) {
-        // Populate the metric's labels map
-        metric.labels[key] = value;
-      }
-    }
-
-    // Assign the metric its value and add it to the family list
-    metric.value = stod(metric_value);
-    metrics.push_back(metric);
-  }
-
-  return metrics;
-}
-
-std::string
-HTTPAPIServer::GenerateRequestClass::ExtractKVMetrics(
-    const std::string& prometheus_metrics, const std::string& orca_type)
-{
-  std::string metric_family = KV_CACHE_BLOCK_METRICS_FAMILY;
-  std::vector<PromMetric> kv_cache_metrics =
-      MetricFamilyExtractor(prometheus_metrics, metric_family);
-
-  double tokens_per_block = -1;
-  double used_blocks = -1;
-  double max_blocks = -1;
-
-  for (const auto& metric : kv_cache_metrics) {
-    if (metric.labels.count(KV_CACHE_BLOCK_TYPE) > 0) {
-      std::string type = metric.labels.at(KV_CACHE_BLOCK_TYPE);
-      if (type == KV_CACHE_BLOCK_TYPE_TOKENS_PER) {
-        tokens_per_block = metric.value;
-      } else if (type == KV_CACHE_BLOCK_TYPE_USED) {
-        used_blocks = metric.value;
-      } else if (type == KV_CACHE_BLOCK_TYPE_MAX) {
-        max_blocks = metric.value;
-      }
-    }
-  }
-
-  // Return early if not all kv metrics are found and set.
-  if (tokens_per_block < 0 || used_blocks < 0 || max_blocks < 0) {
-    LOG_ERROR << "One or more of the kv metrics was not found or invalid.";
-    return "";
-  }
-
-  // Calculate derived metrics
-  double kv_cache_utilization = 0;
-  if (max_blocks > 0) {
-    kv_cache_utilization = used_blocks / max_blocks;
-  }
-  double max_token_capacity = max_blocks * tokens_per_block;
-  
-  std::unordered_map<std::string, double> metrics; // metrics vector to pass down
-  metrics[KV_CACHE_UTIL_KEY] = kv_cache_utilization;
-  metrics[MAX_TOKEN_CAPACITY_KEY] = max_token_capacity;
-
-  return OrcaKVMetricHeader(orca_type, metrics);
-}
-
-std::string
-HTTPAPIServer::GenerateRequestClass::OrcaKVMetricHeader(
-    const std::string& orca_type, std::unordered_map<std::string, double> metrics)
-{
-  // Logic to construct and format response header
-  std::string header_contents = "";
-  const std::string named_metrics_key = NAMED_METRICS;
-  const std::string kv_util_key = KV_CACHE_UTIL_KEY;
-  const std::string max_token_key = MAX_TOKEN_CAPACITY_KEY;
-
-  if (orca_type == "json") {
-    // Format the metrics according to the ORCA protocol as JSON.
-    triton::common::TritonJson::Value orca_metrics(
-        triton::common::TritonJson::ValueType::OBJECT);
-    triton::common::TritonJson::Value named_metrics(
-        orca_metrics, triton::common::TritonJson::ValueType::OBJECT);
-
-    named_metrics.AddDouble(kv_util_key.c_str(), metrics[kv_util_key]);
-    named_metrics.AddUInt(max_token_key.c_str(), metrics[max_token_key]);
-    orca_metrics.Add(named_metrics_key.c_str(), std::move(named_metrics));
-
-    triton::common::TritonJson::WriteBuffer buffer;
-    orca_metrics.Write(&buffer);
-    header_contents = std::string("JSON ") + buffer.Contents();
-
-  } else if (orca_type == "text") {
-    // Format the metrics according to the ORCA protocol as Native HTTP
-    // (comma separated list).
-    const std::string prefix = named_metrics_key + ".";
-
-    header_contents = "TEXT ";
-    header_contents += prefix + kv_util_key + "=" +
-                       std::to_string(metrics[kv_util_key]) + ", ";
-    header_contents +=
-        prefix + max_token_key + "=" + std::to_string(static_cast<uint64_t>(metrics[max_token_key]));
-  } else {
-    LOG_ERROR << "orca_type is set to an invalid type: " << orca_type;
-  }
-
-  return header_contents;
-}
-#endif  // TRITON_ENABLE_METRICS
 
 void
 HTTPAPIServer::GenerateRequestClass::ChunkResponseCallback(
