+++ src/http_server.h
@@ -39,6 +39,7 @@
 
 #include "common.h"
 #include "data_compressor.h"
+#include "orca_http.h"
 #include "restricted_features.h"
 #include "shared_memory_manager.h"
 #include "tracer.h"
@@ -428,12 +429,6 @@ class HTTPAPIServer : public HTTPServer {
     const MappingSchema* ResponseSchema() { return response_schema_; }
 
    private:
-#ifdef TRITON_ENABLE_METRICS
-    struct PromMetric {
-      std::unordered_map<std::string, std::string> labels;
-      double value;
-    };
-#endif  // TRITON_ENABLE_METRICS
     struct TritonOutput {
       enum class Type { RESERVED, TENSOR, PARAMETER };
       TritonOutput(Type t, const std::string& val) : type(t), value(val) {}
@@ -444,22 +439,6 @@ class HTTPAPIServer : public HTTPServer {
       // TENSOR, PARAMETER type
       uint32_t index;
     };
-
-#ifdef TRITON_ENABLE_METRICS
-    // Helper function to get the KV-cache utilization metrics for the
-    // inference response header
-    static std::string ExtractKVMetrics(
-        const std::string& prometheus_metrics, const std::string& orca_type);
-    // Generates a metric struct for a given family with a map of labels and a
-    // value
-    static std::vector<PromMetric> MetricFamilyExtractor(
-        const std::string& input, const std::string& metricFamily);
-    // Creates a header string in the the proper reporting format for provided
-    // KV-cache metrics.
-    static std::string OrcaKVMetricHeader(
-        const std::string& reporting_format,
-        const std::unordered_map<std::string, double> metrics);
-#endif  // TRITON_ENABLE_METRICS
 
     TRITONSERVER_Error* ExactMappingInput(
         const std::string& name, triton::common::TritonJson::Value& value,
